+++ Frontend/App_Start/ServiceBusConfig.cs
@@ -6,6 +6,7 @@ using TaskScheduler.Logging;
 using TaskScheduler.Operations;
 using TaskScheduler.Repositories;
 using TaskScheduler.Services;
+using TaskScheduler.Services.Dates;
 using TaskScheduler.Services.Files;
 
 namespace Frontend
