+++ Frontend/Views/Tasks/_TasksTable.cshtml
@@ -1,5 +1,6 @@
 ﻿@using TaskScheduler
-@model IEnumerable<TaskScheduler.TaskInfo>
+@using TaskScheduler.Entities
+@model IEnumerable<TaskScheduler.Entities.TaskInfo>
 
     <table class="table table-striped">
         <thead>
