+++ TaskScheduler/Entities/TaskConfiguration.cs
@@ -1,0 +1,10 @@
+﻿namespace TaskScheduler.Entities
+{
+    public class TaskConfiguration
+    {
+        public string Name { get; set; }
+        public string CommandType { get; set; }
+        public string CommandParameters { get; set; }
+        public string Frequency { get; set; }
+    }
+}
