+++ TaskScheduler/Entities/TaskStatus.cs
@@ -1,0 +1,8 @@
+﻿namespace TaskScheduler.Entities
+{
+    public enum TaskStatus
+    {
+        Disabled,
+        Enabled
+    }
+}
