+++ TaskScheduler/EventHandlers/UpdateTaskResponseStatusEventHandler.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using TaskScheduler.Entities;
 using TaskScheduler.EventBus;
 using TaskScheduler.Events;
 using TaskScheduler.Repositories;
