--- TaskScheduler/ITaskManager.cs
@@ -1,7 +1,0 @@
-﻿namespace TaskScheduler
-{
-    public interface ITaskManager
-    {
-        void Start();
-    }
-}
