+++ TaskScheduler/Managers/ITaskManager.cs
@@ -1,0 +1,7 @@
+﻿namespace TaskScheduler.Managers
+{
+    public interface ITaskManager
+    {
+        void Start();
+    }
+}
