+++ TaskScheduler/Operations/HttpOperation.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Net;
 using Newtonsoft.Json;
+using TaskScheduler.Entities;
 
 namespace TaskScheduler.Operations
 {
