+++ TaskScheduler/TaskScheduler.csproj
@@ -57,7 +57,7 @@
     <Compile Include="Configuration\Entities\LoggerSettings.cs" />
     <Compile Include="Configuration\Entities\TaskSettings.cs" />
     <Compile Include="Configuration\IConfiguration.cs" />
-    <Compile Include="ErrorEntry.cs" />
+    <Compile Include="Entities\ErrorEntry.cs" />
     <Compile Include="EventBus\Bus.cs" />
     <Compile Include="EventBus\IBus.cs" />
     <Compile Include="EventBus\IEvent.cs" />
@@ -78,7 +78,7 @@
     <Compile Include="Events\UpdateTaskResponseStatusEvent.cs" />
     <Compile Include="Services\Files\FileService.cs" />
     <Compile Include="Services\Files\IFileService.cs" />
-    <Compile Include="Services\IDateService.cs" />
+    <Compile Include="Services\Dates\IDateService.cs" />
     <Compile Include="Repositories\ConfigurationRepository.cs" />
     <Compile Include="Logging\Messages\LogStashLog.cs" />
     <Compile Include="Logging\Messages\TaskStatusChangedLog.cs" />
@@ -89,21 +89,21 @@
     <Compile Include="Operations\EventOperation.cs" />
     <Compile Include="Operations\HttpOperation.cs" />
     <Compile Include="Operations\IOperation.cs" />
-    <Compile Include="ITaskManager.cs" />
+    <Compile Include="Managers\ITaskManager.cs" />
     <Compile Include="Repositories\ITaskRepository.cs" />
     <Compile Include="Operations\IOperationResolver.cs" />
     <Compile Include="Operations\OperationResolver.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Repositories\IConfigurationRepository.cs" />
     <Compile Include="Repositories\IErrorLogRepository.cs" />
-    <Compile Include="ResponseStatus.cs" />
-    <Compile Include="Services\ITaskMonitoringService.cs" />
-    <Compile Include="Services\TaskMonitoringService.cs" />
-    <Compile Include="Services\DateService.cs" />
-    <Compile Include="TaskConfiguration.cs" />
-    <Compile Include="TaskInfo.cs" />
-    <Compile Include="TaskManager.cs" />
-    <Compile Include="TaskStatus.cs" />
+    <Compile Include="Entities\ResponseStatus.cs" />
+    <Compile Include="Services\Tasks\ITaskMonitoringService.cs" />
+    <Compile Include="Services\Tasks\TaskMonitoringService.cs" />
+    <Compile Include="Services\Dates\DateService.cs" />
+    <Compile Include="Entities\TaskConfiguration.cs" />
+    <Compile Include="Entities\TaskInfo.cs" />
+    <Compile Include="Managers\TaskManager.cs" />
+    <Compile Include="Entities\TaskStatus.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
