--- TaskScheduler/TaskStatus.cs
@@ -1,8 +1,0 @@
-﻿namespace TaskScheduler
-{
-    public enum TaskStatus
-    {
-        Disabled,
-        Enabled
-    }
-}
